# BUG: Get common dtype in apply_along_axis