Currently, `np.apply_along_axis` formats the output array according to the first element or array that it processes. However, this can cause truncation of subsequent elements whose `dtype` may be larger than that of the first element.

This commit patches this behaviour by collecting all of the `dtypes` first and then casting at the end according to the `dtype` most amenable to all of them.

Closes #8352.