@mhvk 2016-12-13
I worry this patch, by first creating an object array, and then converting to scalars, will make the common case where the output does have the correct type quite a bit slower. Also, do you have an example where this is an issue? It looks as if the function itself has to give different types of outputs depending on its inputs, which seems strange. If this is really an issue, I would suggest that for these rare cases, one should advise to wrap the function into another one that ensures the output is of the right type.