@gfyoung 2016-12-13
@mhvk : Did you read my PR description?  All of your questions are answered there.  Not to mention, your suggestion would not be feasible because the damage would already have been done in terms of element coercion when calling `apply_along_axis`.

Nevertheless, your point about performance is well taken.  There is a noticeable performance hit:
~~~python
In [1]: import numpy as np
In [2]: arr = np.arange(1000000)
In [3]: %timeit np.apply_along_axis(lambda x: x + 1, 0, arr) 
~~~
`master` : 5.93 ms / loop
`PR` : 82.8 ms / loop
~~~python
In [4]: %timeit np.apply_along_axis(lambda x: [x + 1, 5], 0, arr)
~~~
`master` : 1.88 ms / loop
`PR` : 2.28 ms / loop