@mhvk 2016-12-13
@gfyoung - yes, I did read your description, but as I said I do not understand how one could easily trigger the sitation you are trying to solve, i.e., how one can have a situation where the returned `dtype` will depend on which part of the array one is applying the function to. Could you give an example of a function for which the issue arises?