@mhvk 2016-12-13
Duh...  somehow I missed that link completely. But it does confirm what I felt already: that one is trying to solve corner cases that would be better dealt with in an update to the documentation rather than in a change that causes a significant performance hit to what I would think is the main use case: *numerical* computations.