@gfyoung 2016-12-13
This is not a corner case though, nor do I understand why we should preclude non-numerical computations for this functions.  And if you want a numerical example, here is one:
~~~python
>>> import numpy as np
>>> arr = [1, np.iinfo(np.int64).max]
>>> np.apply_along_axis(lambda x : x + 1, 0, arr)
array([2, -9223372036854775808])  # Forced casting to int64, which is wrong
~~~