@mhvk 2016-12-13
To me, that is just as much a corner case. Anyway, I guess we'll have to agree to disagree, and it is up to the numpy maintainers to decide whether the performance hit is worth it. (Unsubscribing).