@eric-wieser 2016-12-14
Is there any situation this would reasonably arise other than for strings/unicode? Perhaps only those should be special-case?