@gfyoung 2016-12-14
That is a plausible idea.  I did some testing with different numerical types, and everything seems to be okay.  It just seems to the case with strictly strings/unicode.