@shoyer 2016-12-15
I agree with @mhvk that such performance degradation for the typical case is not acceptable. A sane way to handle this would be to explicitly cast to the desired dtype in `func1d`