@gfyoung 2016-12-15
@shoyer : That's not the issue though.  The problem is that we create the output array based on the `dtype` of the result of the first element.  So in the case of strings, `numpy` creates an array of `dtype` "<U19" for example.  Then, if an array length-21 is added, `numpy` simply truncates this element before adding it.