@shoyer 2016-12-15
> The problem is that we create the output array based on the dtype of the result of the first element. 

If the first element of the result has the correct dtype, this isn't a problem, e.g.,
```
>>> arr = np.array([1, np.iinfo(np.int64).max], dtype=object)
>>> np.apply_along_axis(lambda x: x + 1, 0, arr)
array([2, 9223372036854775808L], dtype=object)
```