@eric-wieser 2016-12-15
> If the first element of the result has the correct dtype, this isn't a problem, e.g.,

Seems that adding a `dtype` argument would fix this: