@gfyoung 2016-12-15
@shoyer : Agreed.  What do you think about @eric-wieser proposal?