@eric-wieser 2016-12-15
> If the first element of the result has the correct dtype, this isn't a problem

The issue is that if each element of the result is a string, the length of the string is part of the dtype, and it's pretty unlikely that each string element will be the same length.

Another option would be to addng a `dtype` argument:

    np.apply_along_axis(lambda x: str(x), [1, 2 ,3], dtype=object).astype(str)