@gfyoung 2016-12-15
@shoyer : Agreed.  What do you think about @eric-wieser <a href="https://github.com/numpy/numpy/pull/8363#issuecomment-267401441">proposal</a>?