@shoyer 2016-12-15
I would prefer to fix handling of functions that return 0d strings arrays, rather than adding a separate dtype argument.