@eric-wieser 2016-12-15
That `TypeError` has already been fixed [here](https://github.com/gfyoung/numpy/commit/84b11f57335998e1c2c32e0b1b529ca94895253d#diff-ba13cf8822527da9907ca1822369030dR116), so the `make_object_scalar` version should work on master