@gfyoung 2016-12-17
Rewrote patch to handle special case for `str` type as first element, which dramatically improved performance as seen below:

In [1]: import numpy as np
In [2]: arr = np.arange(1000000)
In [3]: %timeit np.apply_along_axis(lambda x: x + 1, 0, arr) 
master : 5.93 ms / loop
PR : 5.97 ms / loop

In [4]: %timeit np.apply_along_axis(lambda x: [x + 1, 5], 0, arr)
master : 1.88 ms / loop
PR : 1.90 ms / loop