@gfyoung 2016-12-30
Could someone take a look at this?  The changes seem to be both and performant AFAICT.