@homu 2016-12-31
:umbrella: The latest upstream changes (presumably b1b67d6) made this pull request unmergeable. Please resolve the merge conflicts.