@eric-wieser 2017-01-02
So, right now, this fails:

    >>> np.apply_along_axis(lambda r: np.array(np.sum(r)), 0, np.eye(5))
    TypeError: len() of unsized object

Where I would expect an output of `np.ones(5)`.

But this also fails:

    np.apply_along_axis(np.diag, 0, np.eye(5))

Both of these are the case when the result of `f` is an array with `ndim != 1`. I propose we should extend this to work for all ndims