@eric-wieser 2017-01-20
Fun fact: `np.ma.apply_along_axis()` already does what this PR is proposing