@gfyoung 2017-01-21
@eric-wieser : I'm going to close this.  You appear to be moving in a much better direction with this patch than I am given all of these other issues you have uncovered with your patches.  Thanks for all of your feedback!