@eric-wieser 2017-02-13
@gfyoung: That PR re-implementing `apply_along_axis` is now merged, so you might want to take another look at this